--- app/src/main/java/com/udinesfata/expenz/ui/home/tab/HomeContent.kt
@@ -1,8 +1,0 @@
-package com.udinesfata.expenz.ui.home.tab
-
-import androidx.compose.runtime.Composable
-
-@Composable
-fun HomeContent() {
-
-}
