+++ app/src/main/java/com/udinesfata/expenz/ui/home/tab/budget/BudgetContent.kt
@@ -1,4 +1,4 @@
-package com.udinesfata.expenz.ui.home.tab
+package com.udinesfata.expenz.ui.home.tab.budget
 
 import androidx.compose.runtime.Composable
