+++ app/src/main/java/com/udinesfata/expenz/ui/home/tab/dashboard/DashboardContent.kt
@@ -1,0 +1,8 @@
+package com.udinesfata.expenz.ui.home.tab.dashboard
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun DashboardContent() {
+
+}
